refactor(site): unify renderer media-subpath map into one source of truth
Three identical/overlapping Renderer→media-subpath maps had drifted apart
(cdn-code's getMediaSubpath, codegen's getMediaImportSubpath and
getRendererMediaSubpath). Consolidate into a single getMediaSubpath in
types.ts, alongside the renderer taxonomy, so adding a renderer touches one
map instead of three.
